requirements.txt
py-cord
aiohttp
Pillow
openai
python-dotenv
requests
-q -U google-genai
rich
typing 
html2image
google-genai